Work with any offset / any size of image regardless tile size
Because sometimes we may need to cut makeshift / temporary / bad images,
that are not exact tilemaps, and then the dimensions can not be exact.
For dimensions to be exact one would have to pre-edit the image in an image editor....
But since tiling it gives one the opportunity to have the useless tiles trimmed off, this seems an unnecessary step.
So, trim it anyway! Regardless if dimensions are divisible.